[🚚Rename]: 경로 오타 수정 1. molcules -> molecules로 수정
