+++ src/components/organisms/JoinInput/JoinInput.jsx
@@ -1,4 +1,4 @@
-import { InputField } from '@/components/molcules';
+import { InputField } from '@/components/molecules';
 import { useState } from 'react';
 
 const isEmail = (email) => {
